+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 
+* Support for marked-content points and sequences in [HexaPDF::Content::Canvas]
 * Support for property lists in a page's resource dictionary
 
 ## 0.3.0 - 2017-01-25